refactor: remove 'Zen' from signal package type names
Renamed all Zen-prefixed types to avoid confusion with the Zen framework name:
- ComputedZen<T> -> Computed<T>
- ReadonlyZen<T> -> ReadonlySignal<T>
- AnyZen -> AnySignal
- ZenValue -> SignalValue

This makes the signal package naming independent of the framework name,
which is important since @zen/signal is a standalone reactive primitive library
that can be used with or without the @zen/zen framework.

Updated packages:
- @zen/signal: Core type renames
- @zen/zen: Updated imports to use new type names
- @zen/signal-patterns: Updated type references

All packages rebuilt successfully.